Add files via upload
The code has two interfaces, a Terminal-based GUI (User_interface) and a Pytest (test_case):
The user_interface lets a person choose which function they want to run (smallest, second smallest, and nth smallest). The user then has the possibility to create a random array or input the size and each number in the array.
The Pytest case will create a random array and verify if the result from the smallest, second smallest, and nth smallest are correct. The python environment is python 3.6 and higher.

Thank you for giving me this opportunity, if you see anything missing or want me to improve on some aspect of the code please email me at yannickpenda@gmail.com or leave a comment here.

Sincerely,
Yannick Ngole